requirements.txt
python-telegram-bot
python-dotenv
apscheduler
langdetect
pytest